move mergeRules into utils-decorator.ts
